+++ src/backend/mailbox_oauth2/apps.py
@@ -1,0 +1,1 @@
+"""People application to allow OAuth2 authentication with OIDC provider using mailbox."""