+++ CHANGELOG.md
@@ -52,6 +52,7 @@ and this project adheres to
 
 ### Added
 
+- ✨(domains) add periodic tasks to fetch domain status
 - 🧑‍💻(docker) add celery beat to manage periodic tasks
 - ✨(dimail) management command to fetch domain status